modularization: Move first functions to a separate file.
With this, we may, perhaps, achieve some reusability between projects and,
also, ease implementation of unit tests.
