+++ edx_dl/edx_dl.py
@@ -9,51 +9,17 @@ Unix box, in Windows or in Mac OS X.
 """
 
 import argparse
-import cookielib
 import json
 import logging
 import re
-import subprocess
 import sys
 import urllib
 import urllib2
 
 from bs4 import BeautifulSoup
 
-
-EDX_HOMEPAGE = 'https://www.edx.org'
-LOGIN_URL = 'https://www.edx.org/login'
-DASHBOARD = 'https://www.edx.org/dashboard'
-YOUTUBE_VIDEO_ID_LENGTH = 11
-
-
-def get_initial_token():
-    """
-    Create initial connection to get authentication token for future requests.
-
-    Returns a string to be used in subsequent connections with the
-    X-CSRFToken (Cross-Site Request Forgery) header or the empty string if
-    we didn't find any token in the cookies.
-    """
-    cj = cookielib.CookieJar()
-    opener = urllib2.build_opener(urllib2.HTTPCookieProcessor(cj))
-    urllib2.install_opener(opener)
-    opener.open(EDX_HOMEPAGE)
-
-    for cookie in cj:
-        if cookie.name == 'csrftoken':
-            return cookie.value
-
-    return ''
-
-
-def get_page_contents(url, headers):
-    """
-    Get the contents of the page at the URL given by url. While making the
-    request, we use the headers given in the dictionary in headers.
-    """
-    result = urllib2.urlopen(urllib2.Request(url, None, headers))
-    return result.read()
+from utils import EDX_HOMEPAGE, LOGIN_URL, YOUTUBE_VIDEO_ID_LENGTH, \
+    get_initial_token, get_course_list, get_page_contents
 
 
 def parse_args():
@@ -116,7 +82,6 @@ def parse_args():
                         default=False,
                         help='print debugging information')
 
-
     args = parser.parse_args()
 
     # FIXME: check arguments
@@ -136,56 +101,6 @@ def parse_args():
         sys.exit(1)
 
     return args
-
-
-def get_course_list(headers):
-    """
-    Returns a list of tuples with each tuple consisting of:
-
-    * Course name: name of the course we are currently enrolled in.
-    * Course ID: the 'id' of the course.
-    * State: a string saying if the course has started or not.
-    """
-    dash = get_page_contents(DASHBOARD, headers)
-    soup = BeautifulSoup(dash)
-    courses = soup.find_all('article', 'my-course')
-
-    courses_list = []
-    for course in courses:
-        c_name = course.h3.string
-        c_id = course.a['href'].lstrip('/courses/')
-
-        if c_id.endswith('info') or c_id.endswith('info/'):
-            c_id = c_id.rstrip('/info/')
-            state = 'Started'
-        else:
-            c_id = c_id.rstrip('/about/')
-            state = 'Not started'
-
-        courses_list.append((c_id, c_name, state))
-
-    return courses_list
-
-
-def download_videos(video_urls, opts):
-    """
-    Receives a list of URLs given in video_urls and a dictionary called opts
-    with the preferences made by the user and performs the actual download
-    of the videos with URLs with the options given in opts.
-    """
-    # FIXME: Create subdirectories for each lecture and name the files under
-    # those directories
-
-    cmd = ['youtube-dl', '-A']
-
-    if opts.format:
-        cmd.append('-f %d' % opts.format)
-    if opts.subtitles:
-        cmd.append('--write-srt')
-
-    cmd.extend(video_urls)
-
-    subprocess.call(cmd)
 
 
 if __name__ == '__main__':
@@ -215,7 +130,6 @@ if __name__ == '__main__':
     if not resp.get('success', False):
         logging.error('Problems suppling credentials to edX.')
         exit(2)
-
 
     # This doesn't belong here, probably, but in the validation of the
     # command line options, instead.
@@ -253,6 +167,8 @@ if __name__ == '__main__':
 
     # FIXME: Take the week into consideration
     # FIXME: Transform this into a function
+    # FIXME: Consider all courses here.
+    c_id = args.course_id[0]
     logging.info('%s has %d weeks so far.', c_id, len(weeks))
 
     links = [lec_url for week in weeks for lec_url in week[1]]
